+++ CHANGELOG.md
@@ -17,8 +17,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 
+- Remove `Table.auth` and simplify `Table.client`. This functionality is now managed by
+  `bigquery.Client`.
 - In `Table` and `Topic`, the project ID is no longer changed away from what the user provided.
   It was more confusing and dangerous than it was helpful.
+
+### Removed
+
+- `Table.auth`
 
 ## \[v0.3.9\] - 2024-07-02